requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.2
ultralytics==8.0.196
stripe==7.4.0
python-dotenv==1.0.0 